FILE READER BUGFIX
The format of images is different from files, because it calls it the
"URL" of the files. You don't start from the the "C:" section of the
computer, but rather you're confined to the source folder of the game.
Thus, you can't use any images outside or read any images outside of the
program.